Separated node list APIs from of context APIs.
