+++ libraries/vircadia-client/src/client.h
@@ -90,65 +90,4 @@ int vircadia_create_context(vircadia_context_params);
 VIRCADIA_CLIENT_DYN_API
 int vircadia_destroy_context(int id);
 
-/// @brief Connect/jump to server on specified address.
-///
-/// This function only initiates the connection, use
-/// vircadia_connection_status() to check if it was established.
-///
-/// @param id The ID of the context to use.
-/// @param address The address to go to: a "hifi://" address, an IP
-/// address (e.g., "127.0.0.1" or "localhost"), a file:/// address, a
-/// domain name, a named path on a domain (starts with "/"), a position
-/// or position and orientation, or a user (starts with "@").
-///
-/// @return Negative code in case of an error, otherwise 0. \n
-/// Possible error codes: \n
-/// vircadia_error_context_invalid() \n
-/// vircadia_error_context_loss() \n
-VIRCADIA_CLIENT_DYN_API
-int vircadia_connect(int id, const char* address);
-
-/// @brief Retrieve the connection status.
-///
-/// @param id The ID of the context to use.
-/// @return 0 if not connected, 1 if connected, negative error code otherwise. \n
-/// Possible error codes: \n
-/// vircadia_error_context_invalid() \n
-/// vircadia_error_context_loss() \n
-VIRCADIA_CLIENT_DYN_API
-int vircadia_connection_status(int id);
-
-/// @brief Update the internal list of nodes.
-///
-/// @param id The ID of the context to use.
-/// @return Negative code in case of an errors, otherwise 0. \n
-/// Possible error codes: \n
-/// vircadia_error_context_invalid() \n
-/// vircadia_error_context_loss() \n
-VIRCADIA_CLIENT_DYN_API
-int vircadia_update_nodes(int id);
-
-/// @brief Retrieve the count of connected nodes.
-//
-/// The count only changes with explicit call to vircadia_update_nodes().
-///
-/// @param id The ID of the context to use.
-/// @return The count of connected nodes, or a negative error code. \n
-/// Possible error codes: \n
-/// vircadia_error_context_invalid() \n
-/// vircadia_error_context_loss() \n
-VIRCADIA_CLIENT_DYN_API
-int vircadia_node_count(int id);
-
-/// @brief Retrieve the node UUID by index.
-///
-/// The list of nodes only changes with an explicit call to
-/// vircadia_update_nodes(), which may invalidate the pointer returned by this
-/// function.
-///
-/// @param id The ID of the context to use.
-/// @return Pointer to 16 byte UUID or null in case of an error.
-VIRCADIA_CLIENT_DYN_API
-const uint8_t* vircadia_node_uuid(int id, int index);
-
 #endif /* end of include guard */
